iio: support writing processed values to IIO channels
Add a new function to the IIO API which allows consumers to write
processed values to IIO channels.

This is particularly useful for supporting hardware which has
configurable ADC thresholds. A consumer would be able to change an ADC
threshold by providing a processed value instead of calculating a raw
value.

Change-Id: Ifa2b7a530f59d5992526be75f8a96cf99d430d94
Signed-off-by: Nicholas Troast <ntroast@codeaurora.org>
Signed-off-by: Himanshu Agrawal <hagraw@codeaurora.org>