clk: qcom: Add support to be able to slew PLL
Few alpha PLLs would require to be able to slew in the same VCO mode, which
means the PLL would be able to update the new frequency L value without
turning it off. But to be support this feature the PLL needs to calibrated
at the mid of the VCO range and then enabled.

Change-Id: Ifcd12a9dc1169c0f3109908a59a0fad86edfdb70
Signed-off-by: Taniya Das <tdas@codeaurora.org>
Signed-off-by: Himanshu Agrawal <hagraw@codeaurora.org>